test: Calendar 테스트 코드 추가
