+++ backend/src/test/java/com/woowacourse/momo/group/domain/calendar/DeadlineTest.java
@@ -14,28 +14,18 @@ import com.woowacourse.momo.global.exception.exception.MomoException;
 
 class DeadlineTest {
 
-    private static final Duration DURATION = new Duration(LocalDate.now().plusDays(10), LocalDate.now().plusDays(11));
-
     @DisplayName("마감 기한이 현재 시간보다 이전일 경우 예외가 발생한다")
     @Test
     void deadlineMustBeBeforeNow() {
-        assertThatThrownBy(() -> new Deadline(LocalDateTime.now().minusDays(1), DURATION))
+        assertThatThrownBy(() -> new Deadline(LocalDateTime.now().minusDays(1)))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("마감시간이 과거일 수 없습니다.");
-    }
-
-    @DisplayName("마감 기한이 기간의 시작 일자보다 이전일 경우 예외가 발생한다")
-    @Test
-    void deadlineMustBeBeforeStartDuration() {
-        assertThatThrownBy(() -> new Deadline(LocalDateTime.of(DURATION.getStartDate().plusDays(1), LocalTime.now()), DURATION))
-                .isInstanceOf(MomoException.class)
-                .hasMessage("마감시간이 시작 일자 이후일 수 없습니다.");
     }
 
     @DisplayName("마감 기한이 지났는지 확인한다")
     @Test
     void isOver() {
-        Deadline deadline = new Deadline(LocalDateTime.now().plusDays(1), DURATION);
+        Deadline deadline = new Deadline(LocalDateTime.now().plusDays(1));
 
         assertThat(deadline.isOver()).isFalse();
     }
