+++ backend/src/test/java/com/woowacourse/momo/group/domain/group/GroupFindRepositoryTest.java
@@ -200,7 +200,7 @@ class GroupFindRepositoryTest {
 
     private void setPastDeadline(Group group, LocalDateTime date) throws IllegalAccessException {
         LocalDateTime original = LocalDateTime.of(group.getDuration().getStartDate().minusDays(1), LocalTime.now());
-        Deadline deadline = new Deadline(original, group.getDuration());
+        Deadline deadline = new Deadline(original);
         Calendar calendar = new Calendar(group.getSchedules(), group.getDuration(), original);
 
         int index = 0;
